+++ eeschema/widgets/pin_type_combobox.cpp
@@ -65,5 +65,5 @@ ELECTRICAL_PINTYPE PinTypeComboBox::GetPinTypeSelection()
 
 void PinTypeComboBox::SetSelection( ELECTRICAL_PINTYPE aType )
 {
-    wxBitmapComboBox::SetSelection( aType );
+    wxBitmapComboBox::SetSelection( static_cast<int>( aType ) );
 }
